Move notebook visualization to jupyter_magics.py
